feat: CoinPriceListChart 컴포넌트 분리
