--- frontend/src/components/CoinPriceList/CoinPriceList.style.js
@@ -1,8 +1,0 @@
-import Box from '@mui/material/Box';
-import { styled } from '@mui/material/styles';
-
-export const Container = styled.article`
-  background-color: #000;
-  width: 400px;
-  height: 500px;
-`;
