+++ frontend/src/components/CoinPriceListChart/CoinPriceListChart.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import CoinSearchSection from './CoinSearchSection';
+import CoinPriceListTab from './CoinPriceListTab';
+import CoinPriceList from './CoinPriceList';
+
+import {
+  CoinPriceListContainer,
+} from './CoinPriceListChart.style';
+
+function CoinPriceListChart() {
+  return (
+    <>
+      <CoinPriceListContainer>
+        <CoinSearchSection />
+        <CoinPriceListTab />
+        <CoinPriceList />
+      </CoinPriceListContainer>
+    </>
+  );
+}
+
+export default CoinPriceListChart;
