+++ frontend/src/components/index.js
@@ -1,3 +1,3 @@
 export { default as PostCard } from './PostCard/PostCard';
 export { default as MenuBar } from './menuBar/MenuBar';
-export { default as CoinPriceList } from './CoinPriceList/CoinPriceList';
+export { default as CoinPriceListChart } from './CoinPriceListChart/CoinPriceListChart';
